pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "topic-modeling-with-llms"
version = "0.1.0"
description = "End-to-end topic modeling pipeline with LLM post-processing"
requires-python = ">=3.10"
dependencies = [
    "bertopic>=0.16",
    "fastopic>=0.1",
    "sentence-transformers>=2.2",
    "umap-learn>=0.5",
    "hdbscan>=0.8",
    "datasets>=2.14",
    "anthropic>=0.30",
    "optuna>=3.4",
    "gensim>=4.3",
    "pandas>=2.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "scikit-learn>=1.3",
    "numpy>=1.24",
    "tqdm>=4.65",
    "langdetect>=1.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
]
gcp = [
    "google-cloud-storage>=2.10",
    "google-cloud-aiplatform>=1.40",
]

[project.scripts]
run-pipeline = "topic_modeling.pipelines.pipeline:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-v --tb=short"
